vcs, pkg/workspace: remove funcs and tests that where only used by tests
--- sturdy ---
user_id: 847dfd0c-49bf-40c5-8870-74a12fca0d60
view_id: b50b083d-34cd-4f88-b28e-fddbd246ee8c
workspace_id: 0c20dc01-ae20-4c97-9b5e-65f81092dd4d
